refactor(category): remove query manager, unneeded injection tokens, and unneeded interfaces
