--- libs/category/src/drivers/injection-tokens/category-response-transformer.token.ts
@@ -1,10 +1,0 @@
-import { InjectionToken } from '@angular/core';
-
-import { DaffCategoryResponseTransformerInterface } from '../interfaces/category-response-transformer.interface';
-import { DaffGetCategoryResponse } from '../../models/get-category-response';
-
-/**
- * Injection token that serves as a placeholder for any service that implements the DaffCategoryResponseTransformerInterface.
- */
-export const DaffCategoryResponseTransformer = 
-    new InjectionToken<DaffCategoryResponseTransformerInterface<DaffGetCategoryResponse>>('DaffCategoryResponseTransformer');
