--- libs/category/src/drivers/interfaces/category-page-configuration-transformer.interface.ts
@@ -1,5 +1,0 @@
-import { DaffCategoryPageConfigurationState } from '../../models/category-page-configuration-state';
-
-export interface DaffCategoryPageConfigTransformerInterface<T extends DaffCategoryPageConfigurationState> {
-  transform(category: any, sortFields: any): T;
-}
