--- libs/category/src/drivers/interfaces/category-query-manager.interface.ts
@@ -1,9 +1,0 @@
-import { QueryOptions } from 'apollo-client';
-
-export interface DaffCategoryQueryManagerInterface {
-  getACategoryQuery(
-    identifier: string | number,
-    currentPage: number,
-    pageSize: number
-  ): QueryOptions;
-}
