--- libs/category/src/drivers/interfaces/category-response-transformer.interface.ts
@@ -1,5 +1,0 @@
-import { DaffGetCategoryResponse } from '../../models/get-category-response';
-
-export interface DaffCategoryResponseTransformerInterface<T extends DaffGetCategoryResponse> {
-  transform(category: any): T;
-}
