--- libs/category/src/drivers/interfaces/category-transformer.interface.ts
@@ -1,5 +1,0 @@
-import { DaffCategory } from '../../models/category';
-
-export interface DaffCategoryTransformerInterface<T extends DaffCategory> {
-  transform(category: any): T;
-}
