+++ libs/category/src/drivers/magento/category.service.spec.ts
@@ -8,9 +8,9 @@ import { DaffProductFactory } from '@daffodil/product/testing';
 import { DaffCategoryFactory, DaffCategoryPageConfigurationStateFactory } from '@daffodil/category/testing';
 
 import { DaffMagentoCategoryService } from './category.service';
-import { DaffMagentoCategoryGraphQlQueryManagerService } from './queries/category-query-manager.service';
 import { DaffCategory } from '../../models/category';
-import { DaffCategoryTransformer } from '../injection-tokens/category-transformer.token';
+import { DaffMagentoCategoryTransformerService } from './transformers/category-transformer.service';
+import { GetACategoryQuery } from './queries/get-a-category';
 
 // Because ApolloTestingModule doesn't support multiple apollo queries in the same get call, this file is difficult to test.
 // Maybe one of us can make a pull request to apollo-angular if we get the time.
@@ -31,8 +31,6 @@ xdescribe('Driver | Magento | Category | CategoryService', () => {
   const categoryPageConfigTransformService = jasmine.createSpyObj('DaffMagentoCategoryPageConfigTransformerService', ['transform']);
   categoryPageConfigTransformService.transform.and.returnValue(transformedCategoryPageConfigurationState);
 
-  let categoryGraphQlQueryManagerService: DaffMagentoCategoryGraphQlQueryManagerService;
-
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
@@ -40,14 +38,12 @@ xdescribe('Driver | Magento | Category | CategoryService', () => {
       ],
       providers: [
         DaffMagentoCategoryService,
-        { provide: DaffCategoryTransformer, useValue: magentoCategoryResponseTransformerService },
-        DaffMagentoCategoryGraphQlQueryManagerService
+        { provide: DaffMagentoCategoryTransformerService, useValue: magentoCategoryResponseTransformerService }
       ]
     });
 
     categoryService = TestBed.get(DaffMagentoCategoryService);
     controller = TestBed.get(ApolloTestingController);
-    categoryGraphQlQueryManagerService = TestBed.get(DaffMagentoCategoryGraphQlQueryManagerService);
   });
 
   it('should be created', () => {
@@ -99,7 +95,7 @@ xdescribe('Driver | Magento | Category | CategoryService', () => {
         expect(category.products).toEqual(transformedProducts);
       });
       
-      const op = controller.expectOne(categoryGraphQlQueryManagerService.getACategoryQuery(parseInt(stubCategory.id, 10)).query);
+      const op = controller.expectOne(GetACategoryQuery)
 
       op.flush({
         data: response
@@ -111,7 +107,7 @@ xdescribe('Driver | Magento | Category | CategoryService', () => {
         expect(magentoCategoryResponseTransformerService.transform).toHaveBeenCalledWith(response.category);
       });
       
-      const op = controller.expectOne(categoryGraphQlQueryManagerService.getACategoryQuery(parseInt(stubCategory.id, 10)).query);
+      const op = controller.expectOne(GetACategoryQuery);
       expect(op.operation.variables.id).toEqual(parseInt(stubCategory.id, 10));
 
       op.flush({
