--- libs/category/src/drivers/magento/queries/category-query-manager.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { DaffMagentoCategoryGraphQlQueryManagerService } from './category-query-manager.service';
-
-describe('DaffMagentoCategoryGraphQlQueryManagerService', () => {
-    beforeEach(() => TestBed.configureTestingModule({}));
-
-    it('should be created', () => {
-        const service: DaffMagentoCategoryGraphQlQueryManagerService = TestBed.get(DaffMagentoCategoryGraphQlQueryManagerService);
-        expect(service).toBeTruthy();
-    });
-});
